refactor(ui): ♻️ update import paths in components/ui
