+++ apps/nextjs/src/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '~/lib/utils';
+import { cn } from '~/common';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
