+++ apps/nextjs/src/components/ui/hover-card.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '~/lib/utils';
+import { cn } from '~/common';
 import * as HoverCardPrimitive from '@radix-ui/react-hover-card';
 
 const HoverCard = HoverCardPrimitive.Root;
